First pass at appWindow/environment structure
